Rewrite forward url validator based on rfc
