+++ src/main/java/uk/gov/companieshouse/lookup/validation/RelativeUrl.java
@@ -6,7 +6,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Constraint(validatedBy = RelativeUrlValidator.class)
+@Constraint(validatedBy = UrlPathValidator.class)
 @Target({ElementType.METHOD, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface RelativeUrl {
