+++ src/main/java/uk/gov/companieshouse/lookup/validation/UrlPathValidator.java
@@ -1,0 +1,15 @@
+package uk.gov.companieshouse.lookup.validation;
+
+import javax.validation.ConstraintValidator;
+import javax.validation.ConstraintValidatorContext;
+import java.util.regex.Pattern;
+
+public class UrlPathValidator implements ConstraintValidator<RelativeUrl, String> {
+
+    private static final Pattern URL_PATTERN = Pattern.compile("^(?![^:/?#]+:)(?!//[^/?#]*).*$"); //based on https://tools.ietf.org/html/rfc3986#page-51
+
+    @Override
+    public boolean isValid(String value, ConstraintValidatorContext context) {
+        return value != null && URL_PATTERN.matcher(value).matches();
+    }
+}
